backend/pyproject.toml
[tool.poetry]
name = "vandyhacks2023backend"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Quart = "^0.19.3"
hypercorn = "^0.14.4"
aiofiles = "^23.2.1"
azure-cognitiveservices-speech = "^1.32.1"
langchain = "^0.0.314"
python-dotenv = "^1.0.0"
openai = "^0.28.1"
tiktoken = "^0.5.1"
twilio = "^8.10.0"
google-cloud = "^0.34.0"
google-cloud-texttospeech = "^2.14.2"
aiohttp = "^3.8.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
